Resolving additional params when list response

+++ calculator-api/src/main/java/com/calculator/calculatorapi/dto/record/RecordListResponse.java
@@ -11,7 +11,7 @@ import java.util.List;
 public class RecordListResponse {
     int page;
     int totalPages;
-    int totalRecords;
+    long totalRecords;
     String nextPageToken;
     String prevPageToken;
     List<RecordDto> records;
